+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 ## Changes
 
-* None
+* Confirm before deleting RSpecs
+  ([#1691](https://github.com/GENI-NSF/geni-portal/issues/1691))
 
 ## Installation Notes